Changed moved AMRA’s package out of core
