+++ Averaging Measurement Result Analyser/src/main/java/de/uka/ipd/sdq/beagle/analysis/averaging/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This plugin contains a simple Measurement Result Analyser that takes the average of
+ * measurement results.
+ */
+package de.uka.ipd.sdq.beagle.analysis.averaging;
