--- Averaging Measurement Result Analyser/src/main/java/de/uka/ipd/sdq/beagle/core/analysis/package-info.java
@@ -1,7 +1,0 @@
-/**
- * This plugin contains a simple Measurement Result Analyser.
- *
- * @author Roman Langrehr
- *
- */
-package de.uka.ipd.sdq.beagle.core.analysis;
